added rights to blog

+++ routes/blogPost.routes.js
@@ -11,7 +11,7 @@ router.post('/create', auth,async (request, response)=> {
         const user=await User.findOne({_id:request.headers.userid});
         const rights=userRights[user.status];
         if (!rights.canAddBlogPosts) {
-            response.status(403).json({status:'red',message:'Access denied'});
+            response.status(403).json({status:'red',message:'Access denied (no rights)'});
             return;
         }
         const {text,title,createdBy}=request.body;
@@ -25,28 +25,16 @@ router.post('/create', auth,async (request, response)=> {
     }
 })
 
-router.post('/update', auth,async (request, response)=> {
-    try {
-        const {id,title,text}=request.body.post;
-        const editedBy=request.body.editedBy;
-        const post=await Post.findOne({_id:id});
-        for (const key in request.body.post) {
-            if (post[key]!==request.body.post[key]) post[key]=request.body.post[key];
-        }
-        post.dateEdited=new Date();
-        post.editedBy=editedBy;
-        await post.save(); 
-        response.status(201).json(post);
-    } catch (e) {
-        console.log(e.message);
-        response.status(500).json({message:'Something went wrong. Try again later'});
-    }
-})
-
 router.post('/delete', auth,async (request, response)=> {
     try {
+        const user=await User.findOne({_id:request.headers.userid});
+        const rights=userRights[user.status];
         const {id}=request.body;
         const post=await Post.findOne({_id:id});
+        if (user._id.toString()!==post.owner.toString()&&!rights.canModerateBlog) {
+            response.status(403).json({status:'red',message:'Access denied (no rights)'});
+            return;
+        }
         await post.delete();
         response.status(201).json({message:'Post was successfuly deleted'})
     } catch (e) {
@@ -54,6 +42,29 @@ router.post('/delete', auth,async (request, response)=> {
     }
 })
 
+router.post('/update', auth,async (request, response)=> {
+    try {
+        const user=await User.findOne({_id:request.headers.userid});
+        const rights=userRights[user.status];
+        const {id,title,text}=request.body.post;
+        const post=await Post.findOne({_id:id});
+        if (user._id.toString()!==post.owner.toString()&&!rights.canModerateBlog) {
+            response.status(403).json({status:'red',message:'Access denied (no rights)'});
+            return;
+        }
+        const editedBy=user.name;
+        for (const key in request.body.post) {
+            if (post[key]!==request.body.post[key]) post[key]=request.body.post[key];
+        }
+        post.dateEdited=new Date();
+        post.editedBy=user.name;
+        await post.save(); 
+        response.status(201).json(post);
+    } catch (e) {
+        console.log(e.message);
+        response.status(500).json({message:'Something went wrong. Try again later'});
+    }
+})
 
 router.get('/', auth,async (request, response)=> {
     try {
